packages/py-image-processor/pyproject.toml
[project]
name = "py-image-processor"
version = "0.1.0"
description = "Scalable image processing and storage utility"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "Pillow>=10.2.0",
    "boto3>=1.34.0",
]

[tool.uv]
package = true
